requirements.txt
boto3==1.20.3
mygeotab==0.8.6
pandas==1.0.3
psycopg2==2.9.2
psycopg2_binary==2.9.1
python-dotenv==0.19.2
SQLAlchemy==1.3.24
